archive/extracted/fukaya-lab-server-main/OruCa/nfc/requirements.txt
nfcpy
requests
websocket-client